subvox/pyproject.toml
[project]
name = "subvox"
version = "0.1.0"
description = "Short-term memory layer for Cortex - the voice that whispers"
requires-python = ">=3.12"
dependencies = [
    "redis>=5.0",
    "httpx>=0.28",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-distro>=0.48b0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-instrumentation-redis>=0.48b0",
]

[project.scripts]
subvox-stop = "subvox.stop_hook:main"
subvox-prompt = "subvox.prompt_hook:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/subvox"]
